// client_old/src/store/types.ts
export type UserDataAction =
  | "USER_DATA_MODIFY_LOADING"
  | "USER_DATA_MODIFY_SUCCESS"
  | "USER_DATA_MODIFY_ERROR"
  | "USER_DATA_FETCH_LOADING"
  | "USER_DATA_FETCH_SUCCESS"
  | "USER_DATA_FETCH_ERROR";

export type UserProfileAction = "USER_PROFILE_FETCH_LOADING" | "USER_PROFILE_FETCH_SUCCESS" | "USER_PROFILE_FETCH_ERROR";

export type UserAccountAction =
  | "USER_ACCOUNT_TOKEN_CHECK_LOADING"
  | "USER_ACCOUNT_TOKEN_CHECK_SUCCESS"
  | "USER_ACCOUNT_TOKEN_CHECK_ERROR"
  | "USER_ACCOUNT_LOGOUT_SUCCESS"
  | "USER_ACCOUNT_FETCH_LOADING"
  | "USER_ACCOUNT_FETCH_SUCCESS"
  | "USER_ACCOUNT_FETCH_ERROR"
  | "USER_ACCOUNT_AUTH_LOADING"
  | "USER_ACCOUNT_AUTH_SUCCESS"
  | "USER_ACCOUNT_AUTH_ERROR";

export type RespondAction =
  | "RESPOND_MODIFY_LOADING"
  | "RESPOND_MODIFY_SUCCESS"
  | "RESPOND_MODIFY_ERROR"
  | "RESPOND_DELETE_LOADING"
  | "RESPOND_DELETE_SUCCESS"
  | "RESPOND_DELETE_ERROR"
  | "RESPOND_FETCH_LOADING"
  | "RESPOND_FETCH_SUCCESS"
  | "RESPOND_FETCH_ERROR"
  | "RESPOND_ADD_LOADING"
  | "RESPOND_ADD_SUCCESS"
  | "RESPOND_ADD_ERROR";

export type JobAction =
  | "JOB_DELETE_LOADING"
  | "JOB_DELETE_SUCCESS"
  | "JOB_DELETE_ERROR"
  | "JOB_FETCH_LOADING"
  | "JOB_FETCH_SUCCESS"
  | "JOB_FETCH_ERROR"
  | "JOB_ADD_LOADING"
  | "JOB_ADD_SUCCESS"
  | "JOB_ADD_ERROR"
  | "JOB_FETCH_SEARCH_SET";

export type AuthorAction = "AUTHOR_FETCH_LOADING" | "AUTHOR_FETCH_SUCCESS" | "AUTHOR_FETCH_ERROR";

export type ActionStatus = "IDLE" | "LOADING" | "ERROR" | "SUCCESS";

export type ActionType = JobAction | RespondAction | UserProfileAction | UserDataAction | UserAccountAction | AuthorAction;

export interface IReduxAction<T = ActionType, U = any> {
  type: T;
  entityName: string;
  payload?: U;
}